refactor: update train command to use new backend
- Update train command to use new API endpoints
- Refactor code to handle changes in data format
- Change command names for a less confusing UX
